third_party/blink/web_tests/svg/dynamic-updates/SVGFEPointLightElement-dom-z-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'z' attribute of the SVGFEPointLightElement object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pointLight.getAttribute('z') is "100"
PASS pointLight.getAttribute('z') is "30"
PASS successfullyParsed is true

TEST COMPLETE
